FOLSPRINGB-95 message headers parsing bugfix
